Time: 0 ms (100%), Space: 8.2 MB (47.95%) - LeetHub

+++ 0191-number-of-1-bits/0191-number-of-1-bits.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int hammingWeight(int n) {
-        long cnt=0;
+        int cnt=0;
         while(n>0)
         {
             int rem=n%2;
